+++ CHANGELOG.md
@@ -26,6 +26,7 @@ None
 - Fixed color violet on user profile on settings screen
 - Fixed version number on admin more screen to show actual version number
 - Fixed an issue where header title was defaulted to route on admin/users/[id]/edit
+- Fixed pasalubong centers not searchable
 
 ## [1.2.0](https://github.com/niceWizzard/thesis-lakad-rn/compare/1.1.0...1.2.0)