+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed bug where optimize button doesn't optimize the itinerary
-- Fixed coordinates text input too crowded due to accuracy
+- Fixed coordinates text input too crowded due to 
+- Fixed bug where admin cannot update places without an error occuring regarding ALL_LANDMARKS
 
 ### Added
 - Added fullscreen mode in location selection dialog