+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed an issue where the optimized route modal overflows outside the screen for larger itineraries
 
+### Changed
+- Changed the border radius of itinerary stop menu item to fit the rounded style of the app
+
 
 ## [1.3.0](https://github.com/niceWizzard/thesis-lakad-rn/compare/1.2.1...1.3.0)